Refine firm decision boundaries
