+++ src/main/scala/com/boombustgroup/amorfati/agents/firm/FirmStepSemantics.scala
@@ -18,6 +18,14 @@ private[agents] object FirmStepSemantics:
   sealed trait Opening          extends Phase
   sealed trait DecisionSelected extends Phase
   sealed trait PrimaryExecuted  extends Phase
+  sealed trait HiringSignaled   extends Phase
+  sealed trait DebtSettled      extends Phase
+  sealed trait OpeningTraced    extends Phase
+  sealed trait GreenInvested    extends Phase
+  sealed trait CapitalInvested  extends Phase
+  sealed trait DigitalDrifted   extends Phase
+  sealed trait InventorySettled extends Phase
+  sealed trait FdiSettled       extends Phase
   sealed trait Settled          extends Phase
   sealed trait Audited          extends Phase
   sealed trait Closed           extends Phase
@@ -28,21 +36,36 @@ private[agents] object FirmStepSemantics:
 
   private inline def unwrap[A, P <: Phase](staged: At[A, P]): A = staged
 
-  /** Complete opening boundary for processing one firm during one execution
-    * month.
-    */
-  final case class Input(
+  /** Firm-local state and same-month world boundary visible at opening. */
+  final case class OpeningState(
       firm: State,
       financialStocks: FinancialStocks,
       world: World,
       executionMonth: ExecutionMonth,
       operationalSignals: OperationalSignals,
+      corpBondDebt: PLN,
+  )
+
+  /** Credit-pricing and bank-approval boundary used by firm decisions. */
+  final case class CreditSurface(
       lendRate: Rate,
       bankCreditDecision: PLN => CreditDecision,
+  )
+
+  /** Stochastic and peer context for this firm decision. */
+  final case class DecisionEnvironment(
       allFirms: Vector[State],
       rng: RandomStream,
-      corpBondDebt: PLN,
       traceDecision: Boolean,
+  )
+
+  /** Complete opening boundary for processing one firm during one execution
+    * month.
+    */
+  final case class Input(
+      opening: OpeningState,
+      credit: CreditSurface,
+      environment: DecisionEnvironment,
   )
 
   type OpeningInput = At[Input, Opening]
@@ -63,13 +86,45 @@ private[agents] object FirmStepSemantics:
     inline def decisionWithAudit: DecisionWithAudit =
       unwrap(selected)
 
-  type PrimaryExecution = At[Result, PrimaryExecuted]
-  type SettledResult    = At[Result, Settled]
-  type AuditedResult    = At[Result, Audited]
-  type ClosedResult     = At[Result, Closed]
+  type PrimaryExecution    = At[Result, PrimaryExecuted]
+  type HiringSignalUpdate  = At[Result, HiringSignaled]
+  type DebtSettlement      = At[Result, DebtSettled]
+  type OpeningTrace        = At[Result, OpeningTraced]
+  type GreenInvestment     = At[Result, GreenInvested]
+  type CapitalInvestment   = At[Result, CapitalInvested]
+  type DigitalDrift        = At[Result, DigitalDrifted]
+  type InventorySettlement = At[Result, InventorySettled]
+  type FdiSettlement       = At[Result, FdiSettled]
+  type SettledResult       = At[Result, Settled]
+  type AuditedResult       = At[Result, Audited]
+  type ClosedResult        = At[Result, Closed]
 
   inline def primaryExecution(result: Result): PrimaryExecution =
     wrap[Result, PrimaryExecuted](result)
+
+  inline def hiringSignalUpdate(result: Result): HiringSignalUpdate =
+    wrap[Result, HiringSignaled](result)
+
+  inline def debtSettlement(result: Result): DebtSettlement =
+    wrap[Result, DebtSettled](result)
+
+  inline def openingTrace(result: Result): OpeningTrace =
+    wrap[Result, OpeningTraced](result)
+
+  inline def greenInvestment(result: Result): GreenInvestment =
+    wrap[Result, GreenInvested](result)
+
+  inline def capitalInvestment(result: Result): CapitalInvestment =
+    wrap[Result, CapitalInvested](result)
+
+  inline def digitalDrift(result: Result): DigitalDrift =
+    wrap[Result, DigitalDrifted](result)
+
+  inline def inventorySettlement(result: Result): InventorySettlement =
+    wrap[Result, InventorySettled](result)
+
+  inline def fdiSettlement(result: Result): FdiSettlement =
+    wrap[Result, FdiSettled](result)
 
   inline def settledResult(result: Result): SettledResult =
     wrap[Result, Settled](result)
